move cumulatives to constants
